refactor(hub): evict store plumbing from kernel (with-store service, pod-store)
