+++ packages/hub/tsup.config.ts
@@ -21,7 +21,6 @@ import { defineConfig } from 'tsup'
 const ENTRIES = {
   index: 'src/index.ts',
   'i18n/index': 'src/with-shape/i18n/index.ts',
-  'store/index': 'src/kernel/store/index.ts',
   'team/index': 'src/with-party/team/index.ts',
   'session/index': 'src/with-party/session/index.ts',
   'history/index': 'src/with-commit/history/index.ts',
